pyproject.toml
[tool.poetry]
name = "djangotutorial"
version = "0.1.0"
description = ""
authors = ["Ignat Balakirev <ivbignal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
django-split-settings = "^1.1.0"
django-environ = "^0.8.1"
whitenoise = "^6.0.0"
django-cors-headers = "^3.11.0"
drf-yasg = "^1.20.0"
django-filter = "^21.1"
django-debug-toolbar = "^3.2.4"
django-simple-history = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

